<!--
Thanks for contributing to Rails!

Please do not make *Draft* pull requests, as they still send
notifications to everyone watching the Rails repo.

Create a pull request when it is ready for review and feedback
from the Rails team :).

If your pull request affects documentation or any non-code
changes, guidelines for those changes are [available
here](https://edgeguides.rubyonrails.org/contributing_to_ruby_on_rails.html#contributing-to-the-rails-documentation)

About this template

The following template aims to help contributors write a good description for their pull requests.
We'd like you to provide a description of the changes in your pull request (i.e. bugs fixed or features added), motivation behind the changes, and complete the checklist below before opening a pull request.

Feel free to discard it if you need to (e.g. when you just fix a typo). -->

### Motivation / Background

We noted that some of our queries have duplicated/triplicated `... table.id IN (?) ...` clauses, give the size of the lists the qeuries were huge. It caused the queries to get killed.

It happened because the AssociationRelation object is passed through several layers and we apply new scopes to it in order to filter the results. 

<!--
Describe why this Pull Request needs to be merged. What bug have you fixed? What feature have you added? Why is it important?
If you are fixing a specific issue, include "Fixes #ISSUE" (replace with the issue number, remove the quotes) and the issue will be linked to this PR.
-->

This Pull Request has been created because I think it would be easier to merge the predicates as they are added to the where chain. By doing that, the IN predicates will be as short as possible, and it could also identify earlier that the query is unboundable.

### Detail

This Pull Request changes the SQL generated for

```ruby
class Post < ActiveRecord::Base
end

post_scope = Post.where(id: [1,2,3])
  .where(id: [1,2,3])
  .where(id: [3,2,1])
  .where(id: ["3",4,5])

puts post_scope.to_sql
```

Before this PR:
```SQL
SELECT "posts".*
FROM "posts" 
WHERE "posts"."id" IN (1, 2, 3)
AND "posts"."id" IN (1, 2, 3)
AND "posts"."id" IN (3, 2, 1)
AND "posts"."id" IN (3, 4, 5)
```

After this PR:
```SQL
SELECT "posts".* 
FROM "posts"
WHERE "posts"."id" IN (3)
```

### Additional information

<!-- Provide additional information such as benchmarks, reference to other repositories or alternative solutions. -->

Based in the suggested benchmark code, I ran the code below to evaluate the change. The main benefit
is in the generate query size, it decreased ~70-80% for all cases scenario.

```ruby
SCENARIOS = {
  "Two items"        => [1, 2],
  "Mixed type items" => [1, "2", 3],
  "Very Long items"  => (1..5000).to_a
}

SCENARIOS.each_pair do |name, value|
  puts
  puts " #{name} ".center(80, "=")
  puts

  post_scope = Post.where(id: value)

  Benchmark.ips do |x|
    x.report("where") do | times |
      post_scope = post_scope.where(id: value)
    end
  end

  post_scope = Post.where(id: value)
  (1..5).each do
    post_scope = post_scope.where(id: value)
  end

  Benchmark.ips do |x|
    x.report("to_sql") { post_scope.to_sql }
  end

  post_scope = Post.where(id: value)
  (1..5).each do
    post_scope = post_scope.where(id: value)
  end

  puts "SQL String bytesize: #{post_scope.to_sql.bytesize}"
end
```

```
================================== Two items ===================================

Calculating -------------------------------------
               where    217.618Q
       where merging    185.838Q (1.15x slower) i/s -    888.267Q in   4.945791s
Calculating -------------------------------------
              to_sql     15.151M 
        to_sql after     15.242M (same-ish) i/s -     77.496M in   5.084469s

            SQL String bytesize: 193
SQL String bytesize after merge: 58 (69% smaller)

=============================== Mixed type items ===============================

Calculating -------------------------------------
               where    195.613Q 
       where merging    152.459Q (1.22x slower) i/s -    730.128Q in   4.953467s
Calculating -------------------------------------
              to_sql     15.240M i/s -     77.665M in   5.096354s
        to_sql after     14.993M (same-ish) i/s -     75.307M in   5.024512s

            SQL String bytesize: 211
SQL String bytesize after merge: 61 (71% smaller)

=============================== Very Long items ================================

Calculating -------------------------------------
               where    107.767T i/s -    524.517T in   4.998719s
       where merging      2.337T (1.97x slower) i/s -     11.673T in   5.004241s
Calculating -------------------------------------
              to_sql     15.261M i/s -     77.715M in   5.092344s
        to_sql after     15.224M (same-ish) i/s -     77.425M in   5.085806s

            SQL String bytesize: 173515
SQL String bytesize after merge: 28945 (83% smaller)
```

### Checklist

Before submitting the PR make sure the following are checked:

* [ ] This Pull Request is related to one change. Changes that are unrelated should be opened in separate PRs.
* [ ] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [ ] Tests are added or updated if you fix a bug or add a feature.
* [ ] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.